Frankenstein together a basic webpack setup
I used `create-react-app` to make the basic setup and then heavily
modified with a combination of online examples and my own custom
preferences. There's a lot of code that was generated in this
process and likely a lot of code that I don't need or want. It
would be great to clean up all this configuration at some point.