Change default ordering of tweets to use `twitter_tweet_id` field
This resolves a bug where some members were not syncing with the
Twitter API because the API was returning tweets with IDs that
already existed in the database. Theory is that this was the result
of inconsistencies in timestamps or timezones. This resolves that
issue by sending API requests strictly with the largest ID in our
database