Fix error syncing tweets
Because each continuing member of congress still gets a new
Member object when a new session of congress begins, the
sync_tweets command was attempting to sync each member as a new
member from scratch. This created an error because it attempted to
sync some tweets that were already in the database.

This commit hackily fixes that by introducing a method to get the
latest tweet either from the member the method is called from or
from any member that has the same ProPublica ID.

The ultimate fix to this will be to separate the Member model into
Member and Term models and associating tweets with the Member.